Refactor kotlin packages related to web controllers.
