+++ src/main/kotlin/fi/hsl/jore4/mapmatching/controller/api/PublicTransportRouteFindRequestDTO.kt
@@ -1,4 +1,4 @@
-package fi.hsl.jore4.mapmatching.api
+package fi.hsl.jore4.mapmatching.controller.api
 
 import fi.hsl.jore4.mapmatching.model.LatLng
 import jakarta.validation.Valid
