+++ src/main/kotlin/fi/hsl/jore4/mapmatching/util/controller/ParameterUtils.kt
@@ -1,4 +1,4 @@
-package fi.hsl.jore4.mapmatching.web.util
+package fi.hsl.jore4.mapmatching.util.controller
 
 import fi.hsl.jore4.mapmatching.model.LatLng
 import fi.hsl.jore4.mapmatching.model.VehicleMode
